+++ CMakeLists.txt
@@ -10,6 +10,7 @@ set(CROSSPLATFORM_SOURCES "src/common/webview.cc")
 if(WIN32)
 	set(WINDOWS_SOURCES ${WINDOWS_SOURCES}
 		src/windows/com_init_wrapper.cpp
+		src/windows/mswebview_engine.cpp
 		src/windows/webview2_loader/reg_key.cpp
 		src/windows/webview2_loader/com_event_handler.cpp
 		src/windows/webview2_loader/msedge_runtime_loader.cpp
